Fixes a 400 error on some servers
Confirmed issues with some servers, reported by a user.

It seems some hosts will throw a 400 ERROR if you used a different User-Agent so I changed it to make it a bit more like others. Bigger problem was using 'Host' in the header all it has is you own web host but that also cause a issue.. 

This was from the original coder and I dont see the need for it.